Refractoring + fix (maybe) the multiple clocks
